requirement.txt
django==4.1.1
djangorestframework==3.13.1
psycopg2-2.9.3
djangorestframework-simplejwt==5.2.0
